Move Attributes to ForgeGradleExtension
