+++ src/main/java/net/minecraftforge/gradle/MinecraftExtension.java
@@ -8,7 +8,6 @@ import groovy.lang.Closure;
 import org.gradle.api.Action;
 import org.gradle.api.artifacts.dsl.RepositoryHandler;
 import org.gradle.api.artifacts.repositories.MavenArtifactRepository;
-import org.gradle.api.attributes.Attribute;
 
 /// The main extension for ForgeGradle, where the Minecraft dependency resolution takes place.
 ///
@@ -55,49 +54,5 @@ public interface MinecraftExtension extends MinecraftMappingsContainer {
         repositories.remove(mavenizer);
         repositories.addFirst(mavenizer);
         return mavenizer;
-    }
-
-    /**
-     * The attributes object for easy reference.
-     * <pre><code>
-     * dependencies {
-     *     implementation 'com.example:example:1.0' {
-     *         attributes.attribute(minecraft.attributes.os, 'windows')
-     *     }
-     * }
-     * </code></pre>
-     *
-     * @return The attributes object
-     * @see Attributes
-     */
-    Attributes getAttributes();
-
-    /// This interface contains the attributes used by the [Minecraft][MinecraftExtension] extension for resolving the
-    /// Minecraft and deobfuscated dependencies.
-    ///
-    /// @see MinecraftExtension#getAttributes()
-    interface Attributes {
-        /// The operating system of the project's host.
-        ///
-        /// This is used to filter natives from the Minecraft repo.
-        ///
-        /// @return The operating system attribute
-        Attribute<String> getOs();
-
-        /// The requested mappings channel of the project.
-        ///
-        /// This is determined using [MinecraftMappings#getChannel()] via [#getMappings()]
-        ///
-        /// @return The mappings channel attribute
-        /// @see #getMappingsVersion()
-        Attribute<String> getMappingsChannel();
-
-        /// The requested mappings version of the project.
-        ///
-        /// This is determined using [MinecraftMappings#getVersion()] via [#getMappings()]
-        ///
-        /// @return The mappings channel version
-        /// @see #getMappingsChannel()
-        Attribute<String> getMappingsVersion();
     }
 }
