+++ src/main/java/net/minecraftforge/gradle/internal/ForgeGradleExtensionInternal.java
@@ -7,6 +7,7 @@ package net.minecraftforge.gradle.internal;
 import net.minecraftforge.gradle.ForgeGradleExtension;
 import org.gradle.api.Action;
 import org.gradle.api.artifacts.repositories.MavenArtifactRepository;
+import org.gradle.api.attributes.Attribute;
 import org.gradle.api.reflect.HasPublicType;
 import org.gradle.api.reflect.TypeOf;
 
@@ -34,5 +35,29 @@ interface ForgeGradleExtensionInternal extends ForgeGradleExtension, HasPublicTy
     @Override
     default Action<MavenArtifactRepository> getMinecraftLibsMaven() {
         return minecraftLibsMaven;
+    }
+
+    @Override
+    default Attributes getAttributes() {
+        return ForgeGradleExtensionInternal.AttributesInternal.INSTANCE;
+    }
+
+    record AttributesInternal() implements Attributes {
+        static AttributesInternal INSTANCE = new AttributesInternal();
+
+        @Override
+        public Attribute<String> getOs() {
+            return ForgeAttributes.OperatingSystem.ATTRIBUTE;
+        }
+
+        @Override
+        public Attribute<String> getMappingsChannel() {
+            return ForgeAttributes.MappingsChannel.ATTRIBUTE;
+        }
+
+        @Override
+        public Attribute<String> getMappingsVersion() {
+            return ForgeAttributes.MappingsVersion.ATTRIBUTE;
+        }
     }
 }
