+++ src/main/java/net/minecraftforge/gradle/internal/MinecraftExtensionInternal.java
@@ -4,11 +4,10 @@
  */
 package net.minecraftforge.gradle.internal;
 
-import net.minecraftforge.gradle.ClosureOwner;
+import net.minecraftforge.gradle.ForgeGradleExtension;
 import net.minecraftforge.gradle.MinecraftExtension;
 import net.minecraftforge.gradle.MinecraftExtensionForProject;
 import org.gradle.api.artifacts.repositories.MavenArtifactRepository;
-import org.gradle.api.attributes.Attribute;
 import org.gradle.api.file.DirectoryProperty;
 import org.gradle.api.provider.Property;
 import org.gradle.api.reflect.HasPublicType;
@@ -22,34 +21,12 @@ interface MinecraftExtensionInternal extends MinecraftExtension, HasPublicType, 
         return TypeOf.typeOf(MinecraftExtension.class);
     }
 
-    @Override
-    default Attributes getAttributes() {
-        return AttributesInternal.INSTANCE;
-    }
-
     Property<MinecraftMappingsInternal> getMappingsProperty();
 
     DirectoryProperty getMavenizerOutput();
 
-    record AttributesInternal() implements Attributes {
-        static AttributesInternal INSTANCE = new AttributesInternal();
-
-        @Override
-        public Attribute<String> getOs() {
-            return ForgeAttributes.OperatingSystem.ATTRIBUTE;
-        }
-
-        @Override
-        public Attribute<String> getMappingsChannel() {
-            return ForgeAttributes.MappingsChannel.ATTRIBUTE;
-        }
-
-        @Override
-        public Attribute<String> getMappingsVersion() {
-            return ForgeAttributes.MappingsVersion.ATTRIBUTE;
-        }
-    }
-
+    // NOTE: This internal interface does NOT implement MinecraftDependencyInternal as it is not actually a dependency!
+    //       The top-level interface implements MinecraftDependency since it acts as a default for all Minecraft dependencies.
     interface ForProject extends MinecraftExtensionForProject, MinecraftExtensionInternal, HasPublicType, MinecraftAccessTransformersContainerInternal {
         @Override
         default TypeOf<?> getPublicType() {
